// src/core/models/diagnostics/DiagnosticSeverity.ts
/**
 * ============================================================================
 * Honor Pole Innovations
 * HonorPole Platform
 * ----------------------------------------------------------------------------
 * File: DiagnosticSeverity.ts
 *
 * Description:
 * Defines the severity level of a diagnostic event.
 * ============================================================================
 */

export enum DiagnosticSeverity {
    Information = "Information",
    Warning = "Warning",
    Error = "Error",
    Critical = "Critical"
}